Move QML documentation example to QDoc tests
The import and types contained in the example ends up in reference
overviews. This causes confusion among users. The example serves its
purpose better as a QML project to generate documentation from as tests
for QDoc.

Moving the example to tests/auto/qdoc will exclude it from ending up in
the online docs. This will also keep the git history cleaner than
readding the same files as a separate commit later on.

Task-number: QTBUG-43904
Change-Id: I6e080d50319f235916cbd8e1dda3da7dba1dccd1
Reviewed-by: Kai Koehne <kai.koehne@qt.io>
Reviewed-by: Topi Reiniö <topi.reinio@qt.io>