QList/Tree/TableWidgetEdit: use correct default flags
The AbstractItemEditor was using the default flags from an
QListWidgetItem as the default values for QList/Tree/tableWidgetItem
which is wrong because the different classes have different default
flags. This lead to an inconsistent behavior during runtime. E.g. a
QTableWidgetItem was not editable but during runtime it was editable.
This fix adds a new function to retrieve the correct default flags
dependending on the selected widget type.

Change-Id: I96e1dec5a459a4d746b469bd773ddb7526e65357
Reviewed-by: Friedemann Kleint <Friedemann.Kleint@qt.io>