qdoc: Add complete version string to the HTML title
Typically, most developers will have a setup with multiple
versions of Qt installed, such as 5.12.0 and 5.12.1. This
means that there will be several versions of the same
document installed. For example, when a developer tries to
search for 'Item' keyword in the Qt Creator help mode, they
will be presented a list of two topics to choose from; one
each for 5.12.0 and 5.12.1. Unless the HTML document
includes patch version info in the title, it is hard to
differentiate between to two topics in the list.

Task-number: QTBUG-73620
Change-Id: I183bf6476f57d65cf2bcae52289dcb4779a87caf
Reviewed-by: Paul Wicking <paul.wicking@qt.io>