QtHelp: Add move operators to QHelpFilterData, QHelpCollectionDetails
Add move constructors and move assignment operators to QHelpFilterData,
QHelpCollectionDetails. Both classes are implemented in terms of a
QSharedDataPointer, which offers a move constructor and move assignment
operator, so it makes sense to just 'export' this to the class API.

In addition, let the compiler create the implementations of the normal
copy constructor and assignment operator.

Change-Id: Id2d8a86b0b1e0a2e75fec989ae1f5442ead50410
Reviewed-by: Friedemann Kleint <Friedemann.Kleint@qt.io>
Reviewed-by: Jarek Kobus <jaroslaw.kobus@qt.io>