qdoc: Allow replacing file lists on \example pages with a link to a URL
Source and image file lists that are automatically generated at the end
of example pages have outlived their usefulness, at least for Qt docs.
When browsing example documentation offline via Qt Creator's Help, the
user has typically already opened the example project. For online
documentation, a better option is to link to a remote repository where
the user can browse the example's content.

This commit adds a new configuration variable 'url.examples' - if
defined, file listings are omitted and a link is constructed for each
example based on the parameter passed to the \example command and
the 'examplesinstallpath' configuration variable.

Task-number: QTBUG-74391
Change-Id: If1f76ee34c5e50236fae626f251e00bf24c1045d
Reviewed-by: Martin Smith <martin.smith@qt.io>