qdoc: Avoid duplicating version strings in HTML <title> element
Commit 802aa37f introduced full version strings into the titles in
HTML header. However, for pages that already included a prefix
of the version information (e.g. 'Qt 5.13'), this meant that
the version appeared now twice ('Qt 5.13 5.13.0').

This commit checks for the presence of a version prefix in the
title and only appends the full version if one is not found or
does not match with the project version (which for Qt modules is
QT_VERSION).

Fixes: QTBUG-74301
Change-Id: Iceb3cf8f0e0181709e547a87cb6e92a6a13dcfef
Reviewed-by: Venugopal Shivashankar <Venugopal.Shivashankar@qt.io>