qtattributionsscanner: Add "PackageComment" field
Add a new field that can be used for general purpose comments
about the package, not to be included in the generated qdoc
documentation.

JSON does not support comments as a first-class concept. A common
workaround  is to use duplicated key entries, so that the first
one can  be used as a  comment field. Anyhow, this is arguably a
hack and not common knowledge.

The name is inspired by the "PackageComment" field in SPDX. The
new feature should already be made part of 5.12 so we can clean
up version information that recently became more prominently
visible as part of the qdoc title -
see https://doc.qt.io/qt-5/licenses-used-in-qt.html

Change-Id: Icfe843d1d725618114442687222c24122d0dd6b9
Reviewed-by: Edward Welbourne <edward.welbourne@qt.io>