qdoc: Make timestamps in LOG lines opt in
Beginning in 5.13, qdoc prints log lines in the the qdoc warning and error
output that tell the reader what qdoc is doing. These can help the reader
quickly see where qdoc is working when it prints warnings and errors. Each
LOG line also included a timestamp, which was useful for measuring how much
time qdoc was using in each phase of each module.

But these timestamps made diffing sequential log files difficult, so the
timestamps are now removed. They can be enabled for debugging qdoc by
adding -timestamps to the qdoc commandline.

Change-Id: I55ae24e642757eed2cb57028bd22626b6efc82f9
Reviewed-by: Friedemann Kleint <Friedemann.Kleint@qt.io>
Reviewed-by: Martin Smith <martin.smith@qt.io>