Add forward looking after Colon
Colons seen are remembered in order to identify a member or base class
identifier.
In case of declaration of a bit field, the colon should be ignored.
If it is not, unexpected behaviors compromise the resolution of
the context, when the context is not explicit in the tr() function.
To differentiate between colons in a member or base class identifier,
and colons in a bit field, this patch implements a heuristic:
We parse forward to either a '{', or a ';'.
If it is a semicolon we assume it's a bit field, if it is a curved open
bracket we assume the colon is for a member or base class identifier.

Related auto test in case of bitfields declaration added.

Task-number: QTBUG-55479
Change-Id: I2ca1124c192e4ab16fd449afc634a5de5023f481
Reviewed-by: Kai Koehne <kai.koehne@qt.io>