qdoc: Don't report override error when reimp is used
qdoc was reporting an error when the reimplementation of a virtual function
that was documented in the base class was given a qdoc comment in the derived
class that only contained the \reimp command. When the reimp command is used,
the error should not be reported. This change implements that.

Change-Id: I6905e35b4e9c3df12617446c04eb9ea6d1804832
Reviewed-by: Paul Wicking <paul.wicking@qt.io>