qdoc: Turn off clang errors in PCH build without include paths
When the build system can't pass qdoc the include paths to qdoc,
qdoc tries to guess reasonable include paths. There has been one
case running qdoc on macOS for the QtPlatformHeaders module, where
qdoc must try to guess the include paths. The guessed paths are
not sufficient, and clang prints a large number of errors caused
by missing stuff. This change tells clang not to report errors
during the PCH build if qdoc had to guess the include paths.

qdoc reports in its log that it guessed the include paths and
that it turned off the clang error reporting.

Change-Id: I91a4242dcc7d3017d511d969621cc3d673c47963
Reviewed-by: Paul Wicking <paul.wicking@qt.io>