qdoc: Avoid generating invalid html
The test qdoc file and qdocconf file provided with QTBUG-74903 gets qdoc
to generate the html file shown below. The line "<li>Test Guide</li>" is
invalid html. It is meant to be part of the navigation bar. This update
prevents the generation of that line, when the navigationbar Text object
is empty when the invalid line is about to be added.

<?xml version="1.0" encoding="UTF-8"?>
<!DOCTYPE html>
<html lang="en">
<head>
  <meta http-equiv="Content-Type" content="text/html; charset=utf-8" />
<!-- test.qdoc -->
  <title>Test Guide | Testing</title>
</head>
<body>
<li>Test Guide</li>
<div class="sidebar">
<div class="toc">
<h3><a name="toc">Contents</a></h3>

<much of html file deleted>

</body>
</html>

The test input html file actually contained an invalid line that
qdoc generated erroneously. This update just removes that invalid
line.

Task-number: QTBUG-74903
Change-Id: I9e8d742676ac8c7c536c73c8f5b98d27aafbd3c7
Reviewed-by: Martin Smith <martin.smith@qt.io>