Add build information for QDoc on Debian-based Linux distributions
For Debian-based Linux distributions, the installation instructions are
a bit misleading. It is generally not necessary to install a full-fledged
LLVM and the requirements for building and running are slightly different.

This patch makes it clear which packages need to be installed and how
LLVM_INSTALL_DIR needs to be configured in that case.

Task-number: QTBUG-68477
Change-Id: I06397d0d16fb29aba07a399b35b6ae3b32d5c80e
Reviewed-by: Martin Smith <martin.smith@qt.io>
Reviewed-by: Simon Hausmann <simon.hausmann@qt.io>