qmake: replace QLinkedList with std::list
This is in preparation of deprecating QLinkedList.

There is but one substantial change: Instead of copying the linked list,
we move it now, and instead of copying items between these two lists, we
use std::list::splice(), which just relinks nodes. It is a bit surprising
that the comment on the ProValueMapStack suggests references into the
container must remain stable, and then some code changes addresses by
making copies of the elements.

This was probably a bug waiting to manifest itself.

This applies the relevant portions of
qtbase/7cba2acd2cc4b68b5a4de2251ab555adb09bf7e8

Change-Id: Ie3d2ff11a38d82ef22e35402c9755757853c9a7a
Reviewed-by: Jörg Bornemann <joerg.bornemann@qt.io>