tests: remove uses of Q_FOREACH
All pretty straight-forward, except two:

- in tst_QHelpPrjectData::filterSections(), use QCOMPARE() on two
  lists instead of running it by hand. To allow for variation in
  order, sort the list under test, and make sure the reference list is
  sorted, too.

- in tst_QHelpEngineCore::files(), the trailing printing of 'lst' was
  dead code: either the QCOMPARE succeeded, then 'lst' was empty, or
  it failed, in which case QCOMPARE exits the function.  The loop was
  obviously meant to show the remaining items in case the test fails,
  so stuff the loop (w/o Q_FOREACH) into a scope guard.

QtC tricked me into believing that these were the only occurrences in
this module. Alas, it just failed to parse large parts of the module's
pro-files. The module continues to be one of the worst offenders
w.r.t. Q_FOREACH with qdoc and linguist each containing ~150
instances, as well as macdeploy with a another dozen or two. The rest
is clean now.

Change-Id: I89f37ef07ab284da7881d624d9111e89cd2b2cbc
Reviewed-by: Jörg Bornemann <joerg.bornemann@qt.io>