Remove extra \n when reading translator comment from .po file
Extra \n was added at the end of the translator message (cf slurpComment).
This was visible from the ts file in case of convertion.
But the round trip .po to .ts to .po was correct
because of the way messages are read from the .ts file (cf poEscapedLines)

Both functions are modified, no extra \n is added, the roundtrip works,
autoTest are all passed.

Fixes: QTBUG-67908
Change-Id: Ie55572f9609c71c164b2ab939bf0c0d0ca1afc20
Reviewed-by: André Hartmann <aha_1980@gmx.de>
Reviewed-by: Kai Koehne <kai.koehne@qt.io>