qdoc: Ensure Generator::fullDocumentLocation() returns a non-empty string
The function took the parent node's location without checking whether
the parent is the root namespace. For example nodes (and possibly
others), this meant that we got an empty location. This in turn
resulted in QDoc omitting the 'href' attribute when writing an .index
node entry for an example node, and consequently, linking failures.

Fixes: QTBUG-76171
Change-Id: I984ada1b88468aab71d08ba7d102bd8661304dab
Reviewed-by: Martin Smith <martin.smith@qt.io>