qdoc: Build artificial module header
This update lets qdoc build an artificial module header whenever it can't
find the project's module header. The artificial module header contains
a #include "/absolute/path/header.h" for each header found using the
project's headerdirs and headers variables in the qdocconf file. header
files with names ending in _p.h are not included in the artificial module
header. The ordering of the includes in the artificial module header is
not controlled, so the includes must not be order-dependent.

Task-number: QTBUG-76005
Change-Id: I9d0975c571b29e0c3c5972a4e1f78427c595d52d
Reviewed-by: Paul Wicking <paul.wicking@qt.io>