distancefieldgenerator: Avoid exceeding max texture size
Since we use a single area allocator for all textures,
a glyph on the edge between two textures, would end up
spanning two textures. It would be assigned to the first,
and which in turn would be expanded to account for the
additional data. The result was that the texture height
exceeded the maximum size set.

This is actually an inherited problem from Qt Quick, but
it happens more frequently with the distancefieldgenerator
because you will typically generate a large number of
glyphs, and also the default max size set is quite low
to be on the safe side.

The bandaid in this patch is to pad the texture height by
the size of one glyph, which will for the most part get
rid of the problem, except for fonts where a majority of
glyphs exceed the em square.

[ChangeLog][distancefieldgenerator] Fixed a bug where the
generated textures might exceed the maximum height.

Task-number: QTBUG-76528
Change-Id: I51487bbf7c46556b022bfd45cefc8776d0272de3
Reviewed-by: Simon Hausmann <simon.hausmann@qt.io>