distancefieldgenerator: Fix crash with multiple textures
We need to make sure the allocated area starts at (0,0),
since we allocate a texture at [width x height], and this
would be too small if it did not also account for possible
gaps at the top of the cache. This happens because glyphs
may sometimes end up at the edge between two textures, in
which case we will reserve space for them in the second
of the two even though this is not needed.

So this is a deficiency in the cache logic itself, but
we need to at least make sure it does not crash.

[ChangeLog][distancefieldgenerator] Fixed possible crash
when generating large number of glyphs with a small texture
size.

Task-number: QTBUG-76188
Task-number: QTBUG-76528
Change-Id: I84c87c72985f5b621d120e7526cfe071555fa7f8
Reviewed-by: Simon Hausmann <simon.hausmann@qt.io>