qdoc: Revise ClassNode::docMustBeGenerated()
Earlier implementation was incorrect as it omitted generating output
for classes that have only non-function members. Also, it incorrectly
generated output for private, non-documented classes.

Simplify the criteria for generating output; filtering based on the
appearance of 'Private' string is not necessary for Qt, and may
cause problems for projects external to Qt.

Fixes: QTBUG-76892
Change-Id: Ieee6a693f887cbc0244753298f13a9b411801d56
Reviewed-by: Martin Smith <martin.smith@qt.io>