qdoc: Correct algorithm for finding class qualifier
The algorithm for finding the class qualifier for the function name
in a \fn signature was incorrect in cases where the return type also
had a class qualifier. This resulted in some clang error messages
being printed when they should not have been printed because the
class was marked internal. This update corrects that algorithm.

Change-Id: I02983710c73251b8fc75ccb2893ae1d9f5aa0fe6
Reviewed-by: Paul Wicking <paul.wicking@qt.io>