qdoc: Add hasDoc() test to findOverriddenFunction()
When qdoc finds \reimp in a function node comment, it searches the parent class
node's base classes to find the function node for the function that is overridden.
When it finds this node it prints the "Reimpliments: xxx()" line in the current
function node's documentation. But the search for the overridden function is meant
to be recursive. It was working recursively, but the search was ending prematurely
on a function node that had no documentation. In that case, the search should go
higher in the inheritance tree to find an overridden function that is documented.
This update adds the hasDoc() test to that search function to ensure that the final
matching function does have documentation.

Change-Id: Iaf6481f8b3aa98df0d0ef188912f1338316f079a
Reviewed-by: Paul Wicking <paul.wicking@qt.io>