qdoc: Fix the \reimp command output
When qdoc finds \reimp in a function node comment, it searches the parent class
node's base classes to find the function node for the function that is overridden.
When it finds this node it prints the "Reimpliments: Yyy::Xxx()" line in the current
function node's documentation. This was not working correctly when the reimplemented
function was in a different module. The problem was that the fact that the function
in the base class had been documented was not recorded in the index file. This update
fixes that problem. Now every element that was documented is given the "documented"
attribute in the index file set to true.

Sometimes, \reimp is used to say that a function reimplements an access
function for a property in a base class. This update handles that case
as well. See QToolButton::sizeHint(), for example.

This change introduces a 6 new qdoc errors in QtBase about \reimp functions for
which a documented base class function was not found.

Change-Id: Idd444958c3118ade97642bf84781166e6ca8f036
Reviewed-by: Paul Wicking <paul.wicking@qt.io>