qdoc: Improve distribution of shared comment nodes
Until now, QDoc assumed that all nodes sharing a comment must be
function nodes (C++), or grouped properties or methods (QML). This
was a poor assumption, as writers can expect to group together
any other types of documentation, such as macros, typedefs, or QML
signals.

This commit fixes that. Nodes in a shared comment 'collective' are
distributed to correct sections based on the type of the first node
in the collective, so remaining assumption is that all grouped
commands are of the same type.

Fixes: QTBUG-77085
Fixes: QTBUG-77099
Change-Id: I936b24a26443836edb02774a7df6ea6e001a5124
Reviewed-by: Martin Smith <martin.smith@qt.io>
Reviewed-by: Paul Wicking <paul.wicking@qt.io>
Reviewed-by: Mitch Curtis <mitch.curtis@qt.io>