QDoc: Accept command-line arguments from @-files
This enables QDoc to read files given as command-line arguments
that are prepended with '@'. Such a file can include any and all
valid command-line arguments. Specifically, this lets the build
system pass include paths as a file, which may be a requirement
in certain Windows environments due to a platform limitation.

argumentsFromCommandLineAndFile() was shamelessly stolen from moc.
As adding unit tests here basically calls for refactoring the
qdocCommandLineParser class, it was decided to test this in
production. As it works for moc, the assumption is that it will
work also for QDoc.

Task-number: QTBUG-68259
Change-Id: I71698f6514b1343f0a99e7deef82857c1739742e
Reviewed-by: Edward Welbourne <edward.welbourne@qt.io>