qdoc: Fix "Inherited By" output
The "Inherited By" list was incomplete for QObject (and other classes)
because classes that inherited QObject from modules outside QtBase were
not being included. This update cleans up the maintenance of a class's
base class and derived class information so that the "Inherited By" list
is more correct.

Note that the "Inherited By" list can't be guaranteed to be complete
because, for example, classes in external modules might inherit QObject,
but if those external modules are not listed as dependencies of QtCore,
QObject won't know about those classes.

Task-number: QTBUG-77126
Change-Id: Ia0be361f9e909fee04e9680d4219c96812306712
Reviewed-by: Paul Wicking <paul.wicking@qt.io>