QDoc: make the Config a parameter instead of a global
This shall make it easier to test QDoc, by making mocking possible.

This change moves the instantiation of Config:: to reside inside main(),
so that it can be passed as a parameter where it is needed.

This object is then passed to QDocCommandLineParser::process(), so that it
can be changed via the object itself rather than via statics/globals. This
improves the testability of the QDocCommandLineParser class, as the Config
object can easily be mocked.

This change also replaces a couple of foreach() uses in main.cpp with
range-based for.

Task-number: QTBUG-71165
Task-number: QTBUG-71176
Change-Id: I7b98dfc36895cd88fe54dcc9ce31b75aab820b3b
Reviewed-by: Edward Welbourne <edward.welbourne@qt.io>
Reviewed-by: Qt CI Bot <qt_ci_bot@qt-project.org>