QDoc: Further detangle QDocCommandLineParser::process()
This change removes QDocCommandLineParser's dependency on Generator and
Location. The calls to their static methods are moved to a new helper
function in main.cpp, postProcess().
Thus, QDocCommandLineParser no longer sets values in Generator or
Location.

QDir was included via Generator, and is therefore added as an include
in QDocCommandLineParser.

This change further clears up QDocCommandLineParser's external
dependencies, to make it easier to test the class separately.

Task-number: QTBUG-71165
Task-number: QTBUG-71176
Change-Id: I2a024d57d50f84f9271672b6a753e677b748f1b6
Reviewed-by: Martin Smith <martin.smith@qt.io>
Reviewed-by: Edward Welbourne <edward.welbourne@qt.io>