qdoc: Avoid errors related to PCH usage
Remove whitespace from the default parameters passed to Clang, and
add '-ferror-limit=0' to disable the error limit. This helps to
ensure we get a usable precompiled header on all platforms.

Another issue that manifested on Windows was the order in which
the (temporary) module header and the PCH was built - after writing
the module header, its QFile was closed at the end of the function
scope, and it received a timestamp later than the PCH. Some versions
of libclang see this (rightfully) as a problem.

Close the module header file handle before generating the PCH.

Task-number: QTBUG-75053
Change-Id: I61d066c40eddfdfdcc4c8cd847f6bec40652f9e0
Reviewed-by: Paul Wicking <paul.wicking@qt.io>