distancefieldgenerator: Fix garbled text with large fonts on little endian
When writing the texture index of each glyph into the font, we would
first do toBigEndian() on an int and then truncate this to quint16,
causing the texture index to be 0 for all glyphs on little endian
systems.

This would cause glyphs that were not located in the first texture to be
garbled in the output.

[ChangeLog][distancefieldgenerator] Fixed broken text rendering when generating
large glyph sets.

Task-number: QTBUG-77501
Change-Id: I7c2d31a6e57182f440d7f78bd6305109846ccb75
Reviewed-by: Simon Hausmann <simon.hausmann@qt.io>