distancefieldgenerator: Fix performance when selecting unicode ranges
Calling the QItemSelectionModel::select() function that takes a single
index is extremely slow, so for large ranges that span e.g. 20000 glyphs
in the font, such as some of the Chinese ideographic ranges can, the
application would appear to freeze for many seconds while updating
the selections.

Instead we use the overload that takes a QList.

Metrics: For a specific CJK font, this reduces the time taken when
selecting a single unicode range from 30 seconds to 70 ms.

[ChangeLog][Distance Field Generator] Improved performance when
selecting unicode ranges in large fonts.

Task-number: QTBUG-77499
Change-Id: I7ed9bec26b3cbc7e273d305f270a4a6690a81407
Reviewed-by: Kai Koehne <kai.koehne@qt.io>
Reviewed-by: Lars Knoll <lars.knoll@qt.io>