qdoc: Remove superfluous '#' symbol from generated links
This sneaky regression happened as a result of rearranging the
inheritance tree of Node-derived classes.

HtmlGenerator::linkForNode() intends to add #anchor to a link if
it's associated with a Node that does *not* generate an .html
page of its own. It did this by checking whether a page is an
Aggregate. However, since the node tree refactoring, TextPageNode
instances are not Aggregates but do generate .html pages. This
caused excess '#' to be added to all links pointing to a \page.

Luckily, we have Node::isPageNode() which is designed for this
purpose - it tells whether a node has an .html file associated
with it.

Fixes: QTBUG-78040
Change-Id: Ie847c69c4df79075e950aea1ea30c0dcea4f2e4b
Reviewed-by: Paul Wicking <paul.wicking@qt.io>