qdoc: WebXML output: Fix output for \group and \annotatedlist
The WebXMLGenerator did not handle atoms of type AnnotatedList. Adding
this enables listing of grouped pages with the \annotatedlist command.

Similarly, empty output page was generated for the \group page itself.
This was because QDocIndexFiles::generateIndexSections() only handled
groups when processing the root node, which we must skip in this
generator.

Add explicit handling for group nodes, enabling us to receive
callbacks from QDocIndexFiles to add content for these pages.

Task-number: PYSIDE-1088
Change-Id: Id72d589a51fc7aae21ad5cdbb3c1dfa78d4c23ae
Reviewed-by: Venugopal Shivashankar <Venugopal.Shivashankar@qt.io>
Reviewed-by: Paul Wicking <paul.wicking@qt.io>